chore: rename config and add dev script
